+++ compiler/seeds/pthin_skip_tl.from_x.c
@@ -19,7 +19,7 @@ struct parser_asm_token {
   int32_t kind;
   int32_t line;
   int32_t col;
-  int32_t int_val;
+  int64_t int_val;
   double float_val;
   uint8_t *ident;
   int32_t ident_len;
